[ci] Include /bin and /patches in base Dockerfile
This fixes bugs that needed patches to GQLess

+++ bin/postinstall.sh
@@ -1,13 +1,18 @@
 #!/bin/bash
 set -e
 
-cd "$(dirname "$0")/.."
+yarn patch-package
 
-rm -r ./node_modules/react-native || true
-rm -r ./node_modules/react || true
-rm -r ./node_modules/react-dom || true
-rm -r ./node_modules/react-native-svg || true
-ln -s $(realpath ./apps/dish-app/node_modules/react-native) ./node_modules
-ln -s $(realpath ./apps/dish-app/node_modules/react) ./node_modules
-ln -s $(realpath ./apps/dish-app/node_modules/react-dom) ./node_modules
-ln -s $(realpath ./apps/dish-app/node_modules/react-native-svg) ./node_modules
+if [ "$DISH_LINK_RN_MODULES" != "false" ]; then
+  cd "$(dirname "$0")/.."
+  rm -r ./node_modules/react-native || true
+  rm -r ./node_modules/react || true
+  rm -r ./node_modules/react-dom || true
+  rm -r ./node_modules/react-native-svg || true
+  ln -s $(realpath ./apps/dish-app/node_modules/react-native) ./node_modules
+  ln -s $(realpath ./apps/dish-app/node_modules/react) ./node_modules
+  ln -s $(realpath ./apps/dish-app/node_modules/react-dom) ./node_modules
+  ln -s $(realpath ./apps/dish-app/node_modules/react-native-svg) ./node_modules
+  yarn build:refs
+  yarn expo:check-deps
+fi
